renamed core_module as luna_core
